+++ R/preprocess.R
@@ -45,6 +45,10 @@ enw_add_metaobs_features <- function(metaobs, holidays = c(),
   # make day of week a factor
   metaobs[, day_of_week := factor(day_of_week)]
 
+  # add day feature
+  metaobs[, day := as.numeric(date)]
+  metaobs[, day := day - min(day)]
+
   # add week feature
   metaobs[, week := lubridate::week(date)]
   metaobs[, week := week - min(week)]
@@ -682,7 +686,9 @@ enw_construct_data <- function(obs, new_confirm, latest, missing_reference,
 #'
 #' @param max_delay Numeric defaults to 20. The maximum number of days to
 #' include in the delay distribution. Computation scales non-linearly with this
-#' setting so consider what maximum makes sense for your data carefully.
+#' setting so consider what maximum makes sense for your data carefully. Note
+#' that this is zero indexed and so includes the reference date and
+#' `max_delay - 1` other days.
 #'
 #' @param holidays A vector of dates indicating when holidays occur used by
 #' [enw_add_metaobs_features()] to treat holidays as sundays within the
@@ -724,13 +730,6 @@ enw_construct_data <- function(obs, new_confirm, latest, missing_reference,
 #' # Preprocess with default settings
 #' pobs <- enw_preprocess_data(nat_germany_hosp)
 #' pobs
-#'
-#' # Preprocess all data
-#' pobs_all <- enw_preprocess_data(
-#'   germany_covid19_hosp,
-#'   by = c("location", "age_group")
-#' )
-#' pobs_all
 enw_preprocess_data <- function(obs, by = c(), max_delay = 20, holidays = c(),
                                 set_negatives_to_zero = TRUE) {
   obs <- check_dates(obs)