+++ R/utils.R
@@ -146,6 +146,7 @@ utils::globalVariables(
     ".new_group", "observed", "latest_confirm", "mad", "variable",
     "fit", "patterns", ".draws", "prop_reported", "max_confirm",
     "run_time", "cum_prop_reported", "..by_with_group_id",
-    "reference_missing", "prop_missing"
+    "reference_missing", "prop_missing", "day", "posteriors",
+    "formula"
   )
 )