+++ tests/testthat/test-enw_missing_reference.R
@@ -15,8 +15,8 @@ test_that("enw_missing_reference works as expected given correct input data", {
   )
   exp_obs <- data.table::data.table(
     report_date = as.IDate(
-        seq.Date(as.Date("2021-10-01"), as.Date("2021-10-04"), by = "day")
-      ),
+      seq.Date(as.Date("2021-10-01"), as.Date("2021-10-04"), by = "day")
+    ),
     .group = 1, location = "DE",
     confirm = c(rep(0, 3), 4), prop_missing = c(0, NaN, NaN, 1)
   )