+++ tests/testthat/test-enw_model.R
@@ -12,3 +12,9 @@ test_that("enw_model can compile using profiling", {
   skip_on_cran()
   expect_error(enw_model(profile = TRUE), NA)
 })
+
+test_that("enw_model can return stan code without compiling", {
+  skip_on_cran()
+  expect_error(enw_model(compile = FALSE), NA)
+  expect_error(enw_model(compile = FALSE, verbose = FALSE), NA)
+})