+++ NEWS.md
@@ -5,6 +5,7 @@ This is a major release and contains multiple breaking changes. If needing the o
 ## Interface
 
 * A fully featured and flexible formula interface has been added that allows the specification of fixed effects, `lme4` random effects, and random walks. See [#27](https://github.com/epiforecasts/epinowcast/pull/27) by [@seabbs](https://github.com/seabbs).
+* A major overhaul, as described in #57, to the interface of [epinowcast()] with a particular focus on improving the modularity of the model components (described as modules in the documentation). All of the package documentation and vignettes have been updated to reflect this new interface. See #112 by @seabbs.
 
 ## Package
 
@@ -16,6 +17,7 @@ This is a major release and contains multiple breaking changes. If needing the o
 [#106](https://github.com/epiforecasts/epinowcast/pull/106) by [@adrian-lison](https://github.com/adrian-lison) and [@seabbs](https://github.com/seabbs).
 * The discretised reporting probability function has been extended to handle delays beyond the maximum delay in three different ways: ignore, add to maximum, or normalize. The nowcasting model uses "normalise" though work on this is ongoing. See [#113](https://github.com/epiforecasts/epinowcast/pull/113) by [@adrian-lison](https://github.com/adrian-lison) and #121 by @seabbs.
 * Fixed an issue (#105) with `cmdstan 2.30.0` where passing optimisation flags to `stanc_options` by default was causing a compilation error by not passing these flags by default. See #117 by @sbfnk and @seabbs.
+* Addition of regression/integration tests against example data for [epinowcast]() and [enw_preprocess_data()] with convergence checking for several example nowcasting models. Lower level tests for model tools and and model modules have also been added. See #112 by @seabbs.
 
 ## Model